refactor: extract class, DIdConfigMapper
